+++ tests/Rule/ArrayRuleTest.php
@@ -12,7 +12,7 @@ class ArrayRuleTest extends TestCase
     public function testValidateEmptyValue(): void
     {
         // Empty string value. Value is not required by default!
-        $rule = new ArrayRule('name', '');
+        $rule = new ArrayRule('name', []);
         $res = $rule->validate();
         assertThat($res, identicalTo(ArrayRule::VALID));
         assertThat($rule->getValue(), emptyArray());
@@ -46,7 +46,7 @@ class ArrayRuleTest extends TestCase
     public function getArrayValueProvider(): \Generator
     {
         yield 'Given value could be empty' => [
-            '', [], ArrayRule::VALID, ''
+            [], [], ArrayRule::VALID, ''
         ];
 
         yield 'Given value between 4 and 8' => [
@@ -64,6 +64,20 @@ class ArrayRuleTest extends TestCase
         yield 'Given value is not between 4 and 8' => [
             ['Peter', 'Ben', 'Harold'], [ArrayRule::MIN => 4, ArrayRule::MAX => 8], ArrayRule::ERROR,
             "name: The length of array (  0 => 'Peter',  1 => 'Ben',  2 => 'Harold',) is not between 4 and 8"
+        ];
+
+        yield 'Required value should not be an empty array' => [
+            [], [RuleInterface::REQUIRED => true],
+            RuleInterface::ERROR,
+            'name is required and should not be empty: array ()'
+        ];
+
+        yield 'Required value should not be an empty array, with specific message' => [
+            [], [RuleInterface::REQUIRED => true, 'messages' => [
+                RuleInterface::EMPTY_KEY => '%key% is required. `%value%` is empty!'
+            ]],
+            RuleInterface::ERROR,
+            'name is required. `array ()` is empty!'
         ];
     }
 }